fixed logo
