+++ package.json
@@ -23,6 +23,8 @@
   },
   "packageManager": "yarn@1.22.22+sha512.a6b2f7906b721bba3d67d4aff083df04dad64c399707841b7acf00f6b133b7ac24255f2652fa22ae3534329dc6180534e98d17432037ff6fd140556e2bb3137e",
   "devDependencies": {
+    "@babel/core": "^7.26.10",
+    "@babel/preset-env": "^7.26.9",
     "@nuxt/kit": "^3.16.1",
     "@nuxt/types": "^2.18.1",
     "@nuxt/typescript-build": "^3.0.2",
